Pull green bulleted list into shared component file
